1. Removed TransactionCache. We use the HistoryStore now instead. 2. Removed some unnecessary wrapper functions.
